libata: Avoid ABI change in 3.2.62
Commit 1a112d10f03e ('libata: introduce ata_host->n_tags to avoid oops
on SAS controllers') added a field in a structure which may be
driver-allocated.  So we can't simply move it to the end.  However
there is a flags field with many free bits, and we only need 5 bits
(ATA_MAX_QUEUE == 32 and n_tags must be less than this).  Add a
pair of accessors to make this slightly less ugly.


Gbp-Pq: Topic debian
Gbp-Pq: Name libata-avoid-abi-change-in-3.2.62.patch