ARM: dts: sun7i: cubieboard2: add axp209 regulator nodes
This patch adds the regulator nodes for the axp209 by including
the axp209 dtsi. As the inputs of these regulators are from the
axp209's PS output, which is basically just a mux over the 2
inputs, it is considered to be unregulated. Thus we do not provide
input supply properties for them.

The regulator names and constraints are based on the board
schematics and the SoC datasheet.

DCDC2 is used as the cpu power supply. This patch also references
it from the cpu node.

Also get rid of axp209 properties already set in axp209.dtsi.

Signed-off-by: Chen-Yu Tsai <wens@csie.org>
Signed-off-by: Maxime Ripard <maxime.ripard@free-electrons.com>

Conflicts:
	arch/arm/boot/dts/sun7i-a20-cubieboard2.dts

Gbp-Pq: Topic features/arm
Gbp-Pq: Name ARM-dts-sun7i-cubieboard2-add-axp209-regulator-nodes.patch