vfs: Avoid ABI change for dentry union changes
Commit 946e51f2bf37f1656916eb75bd0742ba33983c28 ("move d_rcu from
overlapping d_child to overlapping d_alias") looks disruptive and
it is an API change since the union is named.  However, it doesn't
actually move anything that modules need, so it is not an ABI
change and we can safely hide it from genksysms.

Verify this by adding an unused function with some BUILD_BUG_ONs
to assert the size and alignment of fields remain the same.


Gbp-Pq: Topic debian
Gbp-Pq: Name vfs-avoid-abi-change-for-dentry-union-changes.patch