userns: Fix ABI change in 3.16.7-ckt4
The fix for CVE-2014-8989 added a new member to struct user_namespace.
This is always allocated, and the new member is always used, in
non-modular code.  Move it to the end of the structure and hide it
from genksyms.

Also hide the new #include in kernel/groups.c.


Gbp-Pq: Topic debian
Gbp-Pq: Name userns-fix-abi-change-in-3.16.7-ckt4.patch