mm: Fix pagecache_get_page() ABI change in 3.16.7-ckt6
The last parameter to pagecache_get_page() was dropped, as some
callers passed the wrong value and it is really redundant.

As most filesystems need this function, we should keep the old
function signature for OOT modules.  Rename the function, add a
wrapper with the old name and parameters, and add a macro to make all
new callers use the new function.


Gbp-Pq: Topic debian
Gbp-Pq: Name mm-fix-pagecache_get_page-abi-change-in-3.16.7-ckt6.patch