rtc: hctosys: do not treat lack of RTC device as error
Origin: https://git.kernel.org/linus/aed98b9a1be6fcf1685dfd37f0a3e78e92a21f7d
Bug-Debian: https://bugs.debian.org/784146

When using device trees on the ARM platform, it is not certain at compile
time whether or not the system will have a RTC.

If one enables CONFIG_HCTOSYS just in case the system booted has a RTC,
and it turns out not to be, this will result in a big fat "unable to open
rtc device" error being printed to console, even when "quiet" is set in
the kernel cmdline.

Fix this by outputting the message with loglevel info instead.

Signed-off-by: Floris Bos <bos@je-eigen-domein.nl>
Cc: Alessandro Zummo <a.zummo@towertech.it>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>
Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>

Gbp-Pq: Topic bugfix/all
Gbp-Pq: Name rtc-hctosys-do-not-treat-lack-of-rtc-device-as-error.patch