powerpc/gup: Replace ACCESS_ONCE with READ_ONCE
Forwarded: not-yet

ACCESS_ONCE does not work reliably on non-scalar types. For
example gcc 4.6 and 4.7 might remove the volatile tag for such
accesses during the SRA (scalar replacement of aggregates) step
https://gcc.gnu.org/bugzilla/show_bug.cgi?id=58145)

Change the gup code to replace ACCESS_ONCE with READ_ONCE.

This is not needed upstream as the code has been dropped in
favour of the generic implementation which does not have this
problem.

Signed-off-by: Ben Hutchings <ben@decadent.org.uk>

Gbp-Pq: Topic bugfix/all/access_once
Gbp-Pq: Name 0008-powerpc-gup-Replace-ACCESS_ONCE-with-READ_ONCE.patch