nbd: Add locking for tasks
Origin: http://mid.gmane.org/1444154634-24927-1-git-send-email-mpa@pengutronix.de
Bug-Debian: https://bugs.debian.org/770479

The timeout handling introduced in
	7e2893a16d3e (nbd: Fix timeout detection)
introduces a race condition which may lead to killing of tasks that are
not in nbd context anymore. This was not observed or reproducable yet.

This patch adds locking to critical use of task_recv and task_send to
avoid killing tasks that already left the NBD thread functions. This
lock is only acquired if a timeout occures or the nbd device
starts/stops.

Reported-by: Ben Hutchings <ben@decadent.org.uk>
Signed-off-by: Markus Pargmann <mpa@pengutronix.de>
[bwh: Backported to 3.16: adjust filenames, context]

Gbp-Pq: Topic bugfix/all
Gbp-Pq: Name nbd-add-locking-for-tasks.patch