KEYS: Don't permit request_key() to construct a new keyring
If request_key() is used to find a keyring, only do the search part - don't
do the construction part if the keyring was not found by the search.  We
don't really want keyrings in the negative instantiated state since the
rejected/negative instantiation error value in the payload is unioned with
keyring metadata.

Now the kernel gives an error:

	request_key("keyring", "#selinux,bdekeyring", "keyring", KEY_SPEC_USER_SESSION_KEYRING) = -1 EPERM (Operation not permitted)

Signed-off-by: David Howells <dhowells@redhat.com>
[carnil: Backported for 3.2: adjust context. Search context structure only
introduced in 4bdf0bc300314141e5475e145acb8b5ad846f00d]

Gbp-Pq: Topic bugfix/all
Gbp-Pq: Name KEYS-Don-t-permit-request_key-to-construct-a-new-key.patch