Fix ABI changes for CVE-2013-4312
The fixes for CVE-2013-4312 added new structure members,
user_struct::{unix_inflight,pipe_bufs}.  As this is always allocated in
kernel/user.c and the new member is only used by af_unix which is also
built-in, we can safely add new members at the end.  So move it to the
end and hide it from genksyms.

Similarly for pipe_inode_info::user.


Gbp-Pq: Topic debian
Gbp-Pq: Name fix-abi-changes-for-cve-2013-4312-fix.patch