fs: Fix ABI change for aufs F_SETFL fix
struct file_operations is not embedded in any larger structures so it
is safe to extend it.  Move setfl to the end and hide it from
genksyms.

setfl() still needs to be able to tell whether the
file_operations::setfl pointer is valid, though.  My first thought was
to look at the file's superblock, but that won't work because
e.g. dentries for character devices point to the superblock of the
filesystem where the device node was found!  Instead, add a keyword to
the version string of aufs and make do_setfl() check the name and
version of the module that implements the file_operations.


Gbp-Pq: Topic debian
Gbp-Pq: Name fs-fix-abi-change-for-aufs-f_setfl-fix.patch