[PATCH 7/7] liblockdep: Fix 'defined but not used' warning for init_utsname()
Forwarded: http://mid.gmane.org/20160614204909.GZ7555@decadent.org.uk

We define init_utsname() as static but not inline, resulting
in a warning for every source file that includes lockdep.h but
doesn't call it.

Since it is only used by lockdep.c, define it in there.

Signed-off-by: Ben Hutchings <ben@decadent.org.uk>

Gbp-Pq: Topic bugfix/all
Gbp-Pq: Name liblockdep-fix-defined-but-not-used-warning-for-init.patch