fs: Move procfs/ecryptfs stacking check into ecryptfs
The final upstream fix for CVE-2016-1583 relies on the
super_block::s_stack_depth field which did not exist in 3.2, so that
was added as part of the backport in 3.2.82.  However, that addition
changes ABI.

Revert the changes to add and use s_stack_depth.  Instead, make
ecryptfs specifically prevent mounting on top of procfs, same as it
already did for ecryptfs.

We don't need to touch overlayfs since that doesn't exist here.  We do
have aufs, but that already prevents mounting on top of ecryptfs,
procfs and itself.  It's still possible to mount ecryptfs on top of
aufs, but with only one layer of each, which is consistent with the
upstream restriction to a total of 2 stacked layers.


Gbp-Pq: Topic debian
Gbp-Pq: Name fs-move-procfs-ecryptfs-stacking-check-into-ecryptfs.patch