spi-nor: Add support for n25q256a11 SPI flash device
Forwarded: not-needed
Bug-Debian: https://bugs.debian.org/843650

We have a system that uses the n25q256a11 SPI flash device which is not
marked as supported in the 3.16 kernel series.

This patch is *not* for use in unstable, only Jessie.  A version between 3.16
and 4.7 (stretch) revamped how the chip support is determined, added explicit
support for this chip, and also added a way to add new chips at runtime
without requiring a recompile - all things that make this patch pointless on
unstable/testing.


Gbp-Pq: Topic features/all
Gbp-Pq: Name spi-nor-add-support-for-n25q256a11-spi-flash-device.patch