fs: Fix ABI change in 3.2.84
In order to fix CVE-2015-1350, inode_change_ok() had to be changed
to take a dentry pointer as passed down from inode_setattr(), and
it was renamed to setattr_prepare().

Add inode_change_ok() back, along with a hack to pass the dentry
down via struct attr when there isn't a file pointer there.  In 3.2
there don't appear to be any cases where both the file pointer is
provided and the dentry pointer is needed.  (In upstream there is -
truncate operations set both ATTR_FILE and ATTR_KILL_PRIV.)  WARN
and return an error if that does happen.  


Gbp-Pq: Topic debian
Gbp-Pq: Name fs-fix-abi-change-in-3.2.84.patch