[02/31] net: qcom/emac: add support for the Qualcomm Technologies QDF2400
Origin: https://git.kernel.org/linus/a51f4047232aeb4aee85268b351c7a84b83b36c5

The QDF2432 and the QDF2400 have slightly different internal PHYs,
so there are some programming differences.  Some of the registers in
the QDF2400 have moved, and some registers require different values
during initialization.

Because of the differences, and because HIDs are a scare resource,
the ACPI tables specify the hardware version in an _HRV property.
Version 1 is the QDF2432, and version 2 is the QDF2400.  Any future
SOC that has the same internal PHY but different programming
requirements will be assigned the next available version number.

Signed-off-by: Timur Tabi <timur@codeaurora.org>
Signed-off-by: David S. Miller <davem@davemloft.net>

Gbp-Pq: Topic features/all/qualcomm-emac
Gbp-Pq: Name 0002-net-qcom-emac-add-support-for-the-Qualcomm-Technolog.patch