SCSI: Fix ABI change in 3.16.37
Commit f05795d3d771 ("scsi: Add intermediate STARGET_REMOVE state to
scsi_target_state") added an enumerator to enum scsi_target_state,
and renumbered the existing STARGET_DEL.

The target state doesn't appear to be checked anywhere outside of the
SCSI core, but just in case it is swap STARGET_REMOVE and
STARGET_DEL to maintain the value of the latter.

Hide the change from genksyms.


Gbp-Pq: Topic debian
Gbp-Pq: Name scsi-fix-abi-change-in-3.16.37.patch