[05/31] net: qcom/emac: grab a reference to the phydev on ACPI systems
Origin: https://git.kernel.org/linus/994c5483e7f6dbf9fea622ba2031b9d868feb4b9

Commit 6ffe1c4cd0a7 ("net: qcom/emac: fix of_node and phydev leaks")
fixed the problem with reference leaks on phydev, but the fix is
device-tree specific.  When the driver unloads, the reference is
dropped only on DT systems.

Instead, it's cleaner if up grab an reference on ACPI systems.
When the driver unloads, we can drop the reference without having
to check whether we're on a DT system.

Signed-off-by: Timur Tabi <timur@codeaurora.org>
Reviewed-by: Johan Hovold <johan@kernel.org>
Signed-off-by: David S. Miller <davem@davemloft.net>
[bwh: Adjust to apply after "net: qcom/emac: fix UAF in emac_remove"]

Gbp-Pq: Topic features/all/qualcomm-emac
Gbp-Pq: Name 0005-net-qcom-emac-grab-a-reference-to-the-phydev-on-ACPI.patch